# docs/05-technical/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Travel & Business Trip Management API
  description: |
    API Contract cho hệ thống quản lý công tác doanh nghiệp.
    - **Base URL:** `http://localhost:5000/api/v1`
    - **Auth:** JWT Bearer (Access Token 15m) + Refresh Token (httpOnly cookie, 7d)
    - **Tiền tệ:** `integer` VND nguyên — ví dụ `1500000` = 1.500.000 VNĐ
    - **Tham chiếu:** `architecture.md`, `data-model.md`, `business-rules.md`
    - **FIX-08:** SQLite mutations reserve the database writer before reads. See `concurrency.md`.
      Business audit/notification rows commit atomically; SSE is emitted after commit.
      Contention exhaustion returns 409 CONCURRENT_MODIFICATION; stale state returns 409 INVALID_STATUS_TRANSITION or TRIP_IMMUTABLE.
  version: "1.0.0"
  contact:
    name: Nhóm 11 — MIS3032_1
  license:
    name: Academic Use Only
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development

# ─────────────────────────────────────────────────────────────
# SECURITY SCHEMES
# ─────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token — expires 15 phút.
        Lấy từ `POST /auth/login` hoặc `POST /auth/refresh`.
        Lưu in-memory (Zustand), KHÔNG localStorage.

  # ─────────────────────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────────────────────
  schemas:
...
    TripStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - MANAGER_REVIEWING
        - PENDING_ADMIN_APPROVAL
        - APPROVED
        - ONGOING
        - EXPENSE_DRAFT
        - EXPENSE_SUBMITTED
        - EXPENSE_APPROVED
        - EXPENSE_REJECTED
        - MANAGER_REAPPROVE
        - CLOSED
        - REJECTED
    TimeSlot:
      type: string
      enum: [MORNING, AFTERNOON, EVENING, ALL_DAY]
    ItineraryCategory:
      type: string
      enum: [MEETING, ACCOMMODATION, TRANSPORT, MEAL, OTHER]
...
    ApprovalAction:
      type: string
      enum: [APPROVED, REJECTED]
    ExpenseStatus:
      type: string
      enum: [DRAFT, SUBMITTED, APPROVED, REJECTED, CLOSED]
    ExpenseCategory:
      type: string
      enum: [ACCOMMODATION, TRANSPORT, MEAL, PER_DIEM, OTHER]
...
    ErrorResponse:
      type: object
      required: [error, message, requestId]
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
          description: Machine-readable error code
        message:
          type: string
          example: "departureDate phải là ngày trong tương lai"
        details:
          type: object
          additionalProperties: true
          nullable: true
        requestId:
          type: string
          example: req_abc123

    # ── Policy Check ───────────────────────────────────────────
...
    ItineraryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        itemDate:
          type: string
          format: date
        dayNumber:
          type: integer
          minimum: 1
        timeSlot:
          $ref: '#/components/schemas/TimeSlot'
        location:
          type: string
        activity:
          type: string
        category:
          $ref: '#/components/schemas/ItineraryCategory'
        estimatedCost:
          type: integer
          minimum: 0
          description: VND nguyên
        notes:
          type: string
          nullable: true
        isAiGenerated:
          type: boolean
          readOnly: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ItineraryItemRequest:
      type: object
      required: [itemDate, timeSlot, location, activity, category]
      properties:
        itemDate:
          type: string
          format: date
          description: Phải trong khoảng [departureDate, returnDate] của trip
        timeSlot:
          $ref: '#/components/schemas/TimeSlot'
        location:
          type: string
          minLength: 1
          maxLength: 300
        activity:
          type: string
          minLength: 1
          maxLength: 1000
        category:
          $ref: '#/components/schemas/ItineraryCategory'
        estimatedCost:
          type: integer
          minimum: 0
          default: 0
        notes:
          type: string
          maxLength: 2000
          nullable: true
        sortOrder:
          type: integer
          default: 0

    # ── Expense ────────────────────────────────────────────────
    ExpenseItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        expenseDate:
          type: string
          format: date
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        amount:
          type: integer
          minimum: 1
          description: VND nguyên — phải > 0
        description:
          type: string
        receiptUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ExpenseItemRequest:
      type: object
      required: [expenseDate, category, amount, description]
      properties:
        expenseDate:
          type: string
          format: date
        category:
          $ref: '#/components/schemas/ExpenseCategory'
        amount:
          type: integer
          minimum: 1
          description: VND nguyên — phải > 0
        description:
          type: string
          minLength: 1
          maxLength: 500
        receiptUrl:
          type: string
          maxLength: 500
          nullable: true
    ExpenseClaim:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tripId:
          type: string
          format: uuid
        totalActual:
          type: integer
          readOnly: true
          description: Server-computed = SUM(items.amount)
        estimatedBudgetSnapshot:
          type: integer
          readOnly: true
          description: Snapshot trip.estimatedBudget tại thời điểm tạo expense
        variancePct:
          type: number
          format: float
          nullable: true
          readOnly: true
          description: Server-computed = (totalActual - budgetSnapshot) / budgetSnapshot × 100
        varianceAmount:
          type: integer
          nullable: true
          readOnly: true
          description: Server-computed = totalActual - budgetSnapshot (VND)
        justification:
          type: string
          nullable: true
          description: Bắt buộc khi 0 < variancePct ≤ 10 (BR-TR-05)
        managerReapprovalRequired:
          type: boolean
          readOnly: true
          description: Server-set khi variancePct > 10 (BR-TR-05)
        managerReapproved:
          type: boolean
          readOnly: true
        managerReapproverId:
          type: string
          format: uuid
          nullable: true
          readOnly: true
        managerReapprovedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        status:
          $ref: '#/components/schemas/ExpenseStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseItem'
        submittedAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Notification ───────────────────────────────────────────
...
  responses:
    MutationConflict:
      description: Stale state, CLOSED trip, exhausted concurrency retry or changed payload for an existing Idempotency-Key.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            contention:
              value: { error: CONCURRENT_MODIFICATION, message: 'Dữ liệu đang được cập nhật. Vui lòng thử lại.', details: {}, requestId: req_abc123 }
            idempotency:
              value: { error: IDEMPOTENCY_CONFLICT, message: 'Idempotency-Key đã được dùng với nội dung khác.', details: {}, requestId: req_abc123 }
    Unauthorized:
      description: Token thiếu hoặc không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: UNAUTHORIZED
            message: Token không hợp lệ hoặc đã hết hạn
            requestId: req_abc123
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: "Role 'EMPLOYEE' không có quyền thực hiện action này"
            requestId: req_abc123
    NotFound:
      description: Resource không tồn tại
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Trip không tồn tại
            requestId: req_abc123
    ValidationError:
      description: Input không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: departureDate phải là ngày trong tương lai
            details:
              field: departureDate
              received: "2025-01-01"
            requestId: req_abc123
...
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Same actor/operation/resource/key and payload replay the committed result. Different payload returns 409. Use a new key for a new intent; no content-based item deduplication. Retain the key on retry.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
    tripId:
      name: tripId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    itemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
paths:
...
  /trips/{tripId}/itinerary/{itemId}:
    parameters:
      - $ref: '#/components/parameters/tripId'
      - $ref: '#/components/parameters/itemId'

    patch:
      tags: [Itinerary]
      summary: Cập nhật Itinerary Item
      operationId: updateItineraryItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryItemRequest'
      responses:
        '200':
          description: Item đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

    delete:
      tags: [Itinerary]
      summary: Xóa Itinerary Item
      operationId: deleteItineraryItem
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/expense:
    parameters:
      - $ref: '#/components/parameters/tripId'

    get:
      tags: [Expenses]
      summary: Chi tiết Expense Claim
      operationId: getExpense
      security:
        - BearerAuth: []
      description: |
        **Roles:** Owner (EMPLOYEE), FINANCE.
        `variancePct`, `varianceAmount`, `totalActual`, `estimatedBudgetSnapshot` là read-only.
      responses:
        '200':
          description: Expense Claim với items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseClaim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Expenses]
      summary: Tạo Expense Claim (khởi tạo)
      operationId: createExpense
      security:
        - BearerAuth: []
      description: |
        **Roles:** EMPLOYEE owner.
        **Điều kiện:** trip.status IN (APPROVED, ONGOING) AND chưa có expense.
        Server tự set `estimatedBudgetSnapshot = trip.estimatedBudget`. Không cần request body.
      responses:
        '201':
          description: Expense Claim tạo thành công với status DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseClaim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

    patch:
      tags: [Expenses]
      summary: Cập nhật justification
      operationId: updateExpense
      security:
        - BearerAuth: []
      description: |
        **Roles:** EMPLOYEE owner. Chỉ `justification` được phép cập nhật.
        Mọi computed field bị strip.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                justification:
                  type: string
                  maxLength: 2000
                  nullable: true
      responses:
        '200':
          description: Expense đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseClaim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/expense/submit:
    post:
      tags: [Expenses]
      summary: Nộp Expense Claim
      operationId: submitExpense
      security:
        - BearerAuth: []
      description: |
        **Roles:** EMPLOYEE owner. **Điều kiện:** status = DRAFT AND items ≥ 1.
        **BR-TR-05:**
        - 0 < variance ≤ 10% → cần justification
        - variance > 10% → set managerReapprovalRequired = true
      parameters:
        - $ref: '#/components/parameters/tripId'
      responses:
        '200':
          description: Expense đã submit với variance được tính
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseClaim'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

        '422':
          description: Cần justification (variance 0–10%) hoặc không có items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: JUSTIFICATION_REQUIRED
                message: "Variance 8.24% > 0. Bắt buộc điền lý do vượt chi (BR-TR-05)."
                requestId: req_abc123
  /trips/{tripId}/expense/approve:
    post:
      tags: [Expenses]
      summary: Finance phê duyệt Expense
      operationId: approveExpense
      security:
        - BearerAuth: []
      description: |
        **Roles:** FINANCE.
        **Điều kiện:** expense.status = SUBMITTED AND (nếu managerReapprovalRequired thì managerReapproved = true).
        Response trả về `tripStatus = EXPENSE_APPROVED` để frontend biết enable nút "Đóng hồ sơ".
      parameters:
        - $ref: '#/components/parameters/tripId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  maxLength: 1000
      responses:
        '200':
          description: Expense đã được duyệt — trip chuyển sang EXPENSE_APPROVED
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/ExpenseStatus'
                  approvedAt:
                    type: string
                    format: date-time
                  tripStatus:
                    $ref: '#/components/schemas/TripStatus'
                    description: Giá trị EXPENSE_APPROVED — frontend dùng để enable nút Close
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

        '422':
          description: Variance > 10%, Manager chưa re-approve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /trips/{tripId}/expense/reject:
    post:
      tags: [Expenses]
      summary: Finance từ chối Expense
      operationId: rejectExpense
      security:
        - BearerAuth: []
      description: Employee bổ sung items rồi submit lại.
      parameters:
        - $ref: '#/components/parameters/tripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment:
                  type: string
                  minLength: 10
                  maxLength: 1000
      responses:
        '200':
          description: Expense bị từ chối
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    $ref: '#/components/schemas/ExpenseStatus'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
  /trips/{tripId}/expense/reapprove:
    post:
      tags: [Expenses]
      summary: Manager re-approve phần vượt chi
      operationId: reapproveExpense
      security:
        - BearerAuth: []
      description: |
        **Roles:** MANAGER (manager của employee).
        **BR-TR-05:** Chỉ gọi khi managerReapprovalRequired = true.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
        - $ref: '#/components/parameters/tripId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  $ref: '#/components/schemas/ApprovalAction'
                comment:
                  type: string
                  maxLength: 1000
                  description: Bắt buộc khi action = REJECTED
      responses:
        '200':
          description: Re-approval thực hiện
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  managerReapproved:
                    type: boolean
                  managerReapproverId:
                    type: string
                    format: uuid
                  managerReapprovedAt:
                    type: string
                    format: date-time
                  status:
                    $ref: '#/components/schemas/ExpenseStatus'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
...
  /trips/{tripId}/expense/items/{itemId}:
    parameters:
      - $ref: '#/components/parameters/tripId'
      - $ref: '#/components/parameters/itemId'

    patch:
      tags: [Expense Items]
      summary: Cập nhật khoản chi
      operationId: updateExpenseItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseItemRequest'
      responses:
        '200':
          description: Khoản chi đã cập nhật
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseItem'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'

    delete:
      tags: [Expense Items]
      summary: Xóa khoản chi
      operationId: deleteExpenseItem
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MutationConflict'
...